+++ CanvasPlusPlayground/Features/Files/FoldersPageView.swift
@@ -170,92 +170,6 @@ struct FileRow: View {
     }
 }
 
-@Observable
-class DownloadButtonViewModel {
-    var download: Download?
-
-    init(download: Download?) {
-        self.download = download
-    }
-}
-
-struct DownloadButtonView: View {
-    let model: DownloadButtonViewModel
-
-    var body: some View {
-        DownloadIcon(progress: model.download?.progress, completed: model.download?.localURL != nil)
-    }
-}
-
-struct DownloadIcon: View {
-    var progress: Double?
-    var completed: Bool
-
-    #if os(iOS)
-    var size: CGFloat = 26.0
-    #else
-    var size: CGFloat = 18.0
-    #endif
-
-    var body: some View {
-        ProgressView(value: progress, total: 1.0)
-            .progressViewStyle(GaugeProgressStyle(strokeWidth: size / 12.0))
-            .overlay {
-                Image(systemName: "arrow.down")
-                    .font(.system(size: size * 0.5, weight: .bold))
-                    .offset(x: 0, y: progress == nil ? 0 : size)
-                    .opacity(progress == nil ? 1 : 0)
-                    .foregroundStyle(.secondary)
-            }
-            .overlay {
-                Image(systemName: "checkmark")
-                    .font(.system(size: size * 0.4, weight: .bold))
-                    .offset(x: 0, y: progress == 1 ? 0 : size)
-                    .opacity(progress == 1 ? 1 : 0)
-                    .foregroundColor(.white)
-            }
-            .clipShape(Circle())
-            .animation(.default, value: completed)
-            .animation(.default, value: progress)
-            .frame(width: size, height: size)
-            .font(.system(size: size, weight: .bold))
-    }
-
-    struct GaugeProgressStyle: ProgressViewStyle {
-        var strokeColor = Color.accentColor
-        var strokeWidth = 2.5
-
-        func makeBody(configuration: Configuration) -> some View {
-            let fractionCompleted = configuration.fractionCompleted
-
-            return Circle()
-                .trim(from: 0, to: fractionCompleted ?? 0.0)
-                .stroke(strokeColor, style: StrokeStyle(lineWidth: strokeWidth, lineCap: .round))
-                .rotationEffect(.degrees(-90))
-                .overlay {
-                    if fractionCompleted == 1.0 {
-                        Circle()
-                            .fill(strokeColor)
-                    }
-                }
-                .background {
-                    if fractionCompleted == nil {
-                        Circle()
-                            .stroke(.secondary, style: .init(lineWidth: strokeWidth, lineCap: .butt, lineJoin: .round, dash: [2], dashPhase: 1))
-                    }
-                }
-                .background {
-                    if fractionCompleted != nil {
-                        Circle()
-                            .stroke(.secondary, lineWidth: strokeWidth)
-                    }
-                }
-                .padding(strokeWidth)
-                .animation(.default, value: fractionCompleted)
-        }
-    }
-}
-
 private struct FolderRow: View {
     let folder: Folder
